+++ qase-playwright/changelog.md
@@ -1,3 +1,10 @@
+# playwright-qase-reporter@2.1.6
+
+## What's new
+
+- Added support for status filter in the test run.
+- Improved error handling.
+
 # playwright-qase-reporter@2.1.4
 
 ## What's new